move misplaced tests and hoist inline imports to module level
Move TestBuildSkillCatalog from test_terminal_service_full.py to
test/utils/test_skills.py where the function under test lives. Move
skill utility imports in api/main.py from inline (inside the endpoint
function) to module level for consistency with the rest of the file,
and update test patch targets accordingly.
