+++ test/services/test_terminal_service_full.py
@@ -6,7 +6,6 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from cli_agent_orchestrator.models.agent_profile import AgentProfile
-from cli_agent_orchestrator.models.skill import SkillMetadata
 from cli_agent_orchestrator.models.terminal import TerminalStatus
 from cli_agent_orchestrator.services.terminal_service import (
     OutputMode,
@@ -17,7 +16,6 @@ from cli_agent_orchestrator.services.terminal_service import (
     get_working_directory,
     send_input,
 )
-from cli_agent_orchestrator.utils.skills import build_skill_catalog
 
 
 class TestCreateTerminal:
@@ -281,34 +279,6 @@ class TestCreateTerminal:
         create_terminal(provider_name, "developer", new_session=True)
 
         assert mock_provider_manager.create_provider.call_args.kwargs["skill_prompt"] is None
-
-
-class TestBuildSkillCatalog:
-    """Tests for build_skill_catalog."""
-
-    @patch("cli_agent_orchestrator.utils.skills.list_skills", return_value=[])
-    def test_returns_empty_string_when_no_skills_installed(self, mock_list_skills):
-        """Empty skill stores should produce no injected catalog."""
-        assert build_skill_catalog() == ""
-        mock_list_skills.assert_called_once_with()
-
-    @patch("cli_agent_orchestrator.utils.skills.list_skills")
-    def test_renders_all_installed_skills(self, mock_list_skills):
-        """All installed skills should appear in the global catalog."""
-        mock_list_skills.return_value = [
-            SkillMetadata(name="cao-worker-protocols", description="Worker communication"),
-            SkillMetadata(name="python-testing", description="Pytest conventions"),
-        ]
-
-        assert build_skill_catalog() == (
-            "## Available Skills\n\n"
-            "The following skills are available exclusively in this CAO orchestration context. "
-            "To load a skill's full content, use the `get_skill` MCP tool provided by the "
-            "CAO MCP server. These skills are not accessible through provider-native skill "
-            "commands or directories.\n\n"
-            "- **cao-worker-protocols**: Worker communication\n"
-            "- **python-testing**: Pytest conventions"
-        )
 
 
 class TestGetTerminal:
